# apps/aurelius-rest/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Aurelius REST API",
        "contact": {}
    },
    "host": "localhost:8080",
    "paths": {
        "/agents": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agents"
                ],
                "summary": "Get all agents",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.AgentDto"
                            }
                        }
                    }
                }
            }
        },
        "/jobs": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Get all scheduled jobs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.JobDto"
                            }
                        }
                    }
                }
            }
        },
        "/jobs/{id}": {
            "patch": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Update a scheduled job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Job input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.JobInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job updated successfully",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.AgentDto": {
            "type": "object",
            "properties": {
                "agent_stack": {
                    "type": "string"
                },
                "agent_version": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_seen_at": {
                    "type": "string"
                },
                "registered_at": {
                    "type": "string"
                },
                "service": {
                    "type": "string"
                }
            }
        },
        "dto.JobDto": {
            "type": "object",
            "properties": {
                "agentId": {
                    "type": "string"
                },
                "finishedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "startedAt": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "dto.JobInput": {
            "type": "object",
            "properties": {
                "cron": {
                    "type": "string"
                }
            }
        }
    }
}